extract async operation lookup routes
